smiles2spec_v2/pyproject.toml
[tool.poetry]
name = "smiles2spec-pipeline"
version = "0.1.0"
description = "SMILES to Mass Spectrum Prediction Pipeline"
authors = ["Yusuf Mohamed"]
license = "MIT"
readme = "README.md"
packages = [{include = "smiles2spec", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
numpy = ">=1.24.0,<2.0.0"
pandas = ">=2.0.0"
scikit-learn = ">=1.3.0"
torch = ">=2.0.0"
rdkit = ">=2024.3.1"
pyyaml = ">=6.0"
tqdm = ">=4.65.0"
joblib = ">=1.3.0"
matplotlib = ">=3.7.0"
seaborn = ">=0.12.0"
click = ">=8.1.0"
scipy = ">=1.11.0"
pyarrow = ">=12.0.0"

[tool.poetry.group.dev.dependencies]
pytest = ">=7.0.0"
pytest-cov = ">=4.0.0"
ruff = ">=0.1.0"
ipython = ">=8.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"
